add product page into LastScreen
